field changes to models;

+++ trade_app/models.py
@@ -4,7 +4,7 @@ from django.contrib.auth.models import User
 # Create your models here.
 class Account(models.Model):
     user = models.ForeignKey(User, on_delete=models.CASCADE)
-    cash_balance = models.FloatField(default=1500)
+    cash_balance = models.DecimalField(max_digits=8, decimal_places=2, default=1500.00)
 
     def __str__(self):
         return self.user.username
@@ -31,7 +31,7 @@ class Order(models.Model):
     symbol = models.CharField(max_length=8)
     shares = models.IntegerField(default=0)
     date = models.DateField('purchase date')
-    paid = models.FloatField()
+    paid = models.DecimalField(max_digits=8, decimal_places=2)
 
     def __str__(self):
         return "%s, %d" %  (self.symbol, self.shares)
